// apps/organization/src/app/users/enums/PeopleEnums.ts
export enum PeopleEnum {
  DEPARTMENT = 'Department',
  CLIENT = 'Client',
  PROJECT = 'Project',
  IMAGE = 'Drag & Drop here Or Browse Image',
  PREVIOUS = 'Previous',
  NEXT = 'Next',
  SHOWING = 'Showing',
  ENTRIES = 'entries',
  OF = 'of',
  TO = 'to',
  PAGE_SIZE = 'Page Size',
  NEWPEOPLE = 'New People',
  ERROR_MSG = 'Error fetching User data:',
  LOADING = 'Loading...',
  PEOPLE = 'People',
}
export enum AddPeople {
  FIRST_NAME = 'First Name',
  LAST_NAME = 'Last Name',
  MEMBER_ID = 'Member ID',
  DOB = 'DOB',
  EmployeeID = 'Employee ID',
  companyEmail = 'Company Email Address',
  workPhone = 'Work Phone',
  Phone = 'Personal Phone',
  REPORTING_MANAGER = 'Reporting Manager',
  LOCATION = 'Location',
  BACKUP_MANAGER = 'Backup Manager',
  VIEW_PERMESSIONS = 'View Permissions',
  HIDE_PERMESSIONS = 'Hide Permissions',
  USER_NAME = 'Kevin Peter',
  FULL_NAME = 'FULL NAME',
  EMAIL = 'EMAIL',
  CONTACT = 'CONTACT',
  EMPLOYEE_TYPE = 'EMPLOYEE TYPE',
  WORK_TYPE = 'WORK TYPE',
  SHIFT = 'Shift',
  ROLE = 'Job Role',
  CONTRACT_TYPE = 'Contract Type',
  LAST_LOGIN = 'LAST LOGIN=',
  LAST_UPDATED = 'LAST UPDATED',
  SCREEN = 'Screen',
  PERMESSIONS = 'Permessions',
  SAVE = 'Save',
  CANCEL = 'Cancel',
  EDIT = 'Update',
  New_People = 'Add New People',
  Update_People = 'Edit people',
  BACK_BUTTON = 'Back',
  SELECT_IMG = 'Select New Image',
  EFFECTIVE_PERMESSIONS = 'Effective Permission',
  DROP_IMAGE = 'Drag & Drop here Or Browse',
  LOADING = 'Loading...',
  ERROR_MSG = 'Please select role to view user permessions',
  READ = 'Read',
  WRITE = 'Write',
}

export enum PeopleDetails {
  USER_NAME = 'Kevin Peter',
  FULL_NAME = 'FULL NAME',
  EMAIL = 'EMAIL',
  CONTACT = 'CONTACT',
  EMPLOYEE_TYPE = 'EMPLOYEE TYPE',
  WORK_TYPE = 'WORK TYPE',
  ROLE = 'ROLE',
  LAST_LOGIN = 'LAST LOGIN',
  LAST_UPDATED = 'LAST UPDATED',
  LOADING = 'Loading...',
  ERROR_MSG = 'Error fetching User data:',
  BACK = 'Back',
}
export enum Tablehead {
  NODATA = 'No records...',
  NAME = 'Name',
  EMAIL = 'Email',
  JOBE_ROLE = 'Job Role',
  REPORTING_MANGER = 'Reporting Manager',
  LAST_LOGIN = 'Last Login',
  LAST_UPDATE = 'Last Updated',
  ACTIONS = 'Action',
}
export enum Permessions {
  PERMESSIONS = 'Permessions',
  ACTION = 'Action',
}
export enum message {
  UNARCHIVE = 'Do you really want to un-archive this user?',

  DELETE = 'Do you really want to delete this user?',

  ARCHIVE = 'Do you really want to archive this user?',
}